+++ .gitignore
@@ -12,6 +12,10 @@ dist
 dist-ssr
 *.local
 
+# locales
+src/locales/**/*.ts
+src/locales/**/*.js
+
 # Editor directories and files
 .vscode/*
 !.vscode/extensions.json
