+++ src/helpers/resolveRainbowKitLocale.ts
@@ -1,7 +1,7 @@
 import type { Locale as RainbowKitLocale } from '@rainbow-me/rainbowkit'
 
 import { createLookupTableResolver } from './createLookupTableResolver.ts'
-import { Locale } from '../types'
+import { Locale } from '../types/enum.ts'
 
 export const resolveRainbowKitLocale = createLookupTableResolver<Locale, RainbowKitLocale>(
   {
