--- src/locales/en.js
@@ -1,1 +1,0 @@
-/*eslint-disable*/module.exports={messages:JSON.parse("{\"p6UGtk\":\"Test Button\",\"mnIQMv\":\"Test Text\"}")};
