--- src/locales/en.ts
@@ -1,1 +1,0 @@
-/*eslint-disable*/import type{Messages}from"@lingui/core";export const messages:Messages=JSON.parse("{\"mGQEsL\":[\"<0>+\",[\"0\"],\"</0>Points/h\"],\"i4cwHz\":[\"1 staked MASK will generate \",[\"ratio\"],\" point per hour.\"],\"uyJsf6\":\"About\",\"F0ggQX\":\"About MASK Stake\",\"5MojdE\":\"Accept cookies\",\"hehnjM\":\"Amount\",\"70wH1Q\":\"APR\",\"kfcRb0\":\"Avatar\",\"VuFd5C\":[\"Balance: \",[\"0\"]],\"V/q0wz\":\"Blacklist is powered by <0>Go+</0>\",\"waT/Mj\":\"Can not get determination MASK token\",\"dEgA5A\":\"Cancel\",\"iKvAeG\":\"Checking\",\"hom7qf\":\"Claim\",\"3gad8u\":\"Claim Successfully\",\"7VpPHA\":\"Confirm\",\"jh0i4n\":\"Confirm this transaction in your wallet.\",\"iSLIjg\":\"Connect\",\"HtCmVk\":\"Connect to stake\",\"CtByM7\":\"Connect Wallet\",\"YcfUZ9\":\"Contact Us\",\"AodcaS\":\"Cookie Policy\",\"QSMBSI\":\"Cookies are small text files that are stored on your computer or mobile device when you visit a website. They are used to store information about you and your internet usage habits to improve your online experience. Cookies enable websites to recognize your device and remember some information about your preferences.\",\"DPfwMq\":\"Done\",\"ePK91l\":\"Edit\",\"237hSL\":\"Ended\",\"SlfejT\":\"Error\",\"aqK0+O\":\"Estimated Rewards\",\"aJ4pMe\":\"FAQs\",\"INa0YU\":\"How to Control Cookies\",\"5x0Zyr\":\"If you have any questions or concerns about this Cookie Policy, please contact us at:\",\"v5W7I0\":\"info@dimension.im\",\"znqB4T\":\"Insufficient balance\",\"isuyW5\":\"Introduction\\nMask Network is committed to protecting the privacy of its users. This Privacy Policy outlines how we collect, use, disclose, and safeguard your information when you stake Mask to receive other token rewards through our website. By using our service, you consent to the data practices described in this policy.\\nInformation Collection and Use\\nInformation You Provide to Us\\nWe collect information you provide directly to us when you use our services. This may include:\\nContact information, such as your name, email address, and phone number.\\nBlockchain-related information, such as your wallet addresses.\\nSocial media account information, including but not limited to accounts on platforms like x.com.\\nWhen you choose to link your social media accounts or provide your wallet addresses, we may display this information on our website to enhance the user experience and foster a community environment. We encourage you to review the privacy settings and policies of any social media platforms you interact with to understand how your information may be shared and used.\\nInformation We Collect Automatically\\nWhen you access our website, we may automatically collect information about your device and usage, including:\\nIP address, browser type, and operating system.\\nInformation about your interactions with our website, such as the pages you visit.\\nUse of Your Information\\nWe use the information we collect about you to:\\nProvide, maintain, and improve our services.\\nDisplay your social media profiles and wallet addresses on our website, subject to your preferences and permissions.\\nRespond to your inquiries and provide customer support.\\nMonitor and analyze trends, usage, and activities in connection with our services.\\nSharing of Your Information\\nWe may share your information with third parties in the following situations:\\nWith your consent or at your direction.\\nWith service providers who assist us in our operations, under confidentiality agreements.\\nIf required by law or to protect the rights, property, or safety of [Your Website Name], our users, or others.\\nSecurity\\nWe take reasonable measures to protect your information from unauthorized access, use, alteration, and destruction. However, no internet or email transmission is ever fully secure or error-free.\\nChanges to This Privacy Policy\\nWe may update this Privacy Policy from time to time. We will notify you of any changes by posting the new policy on this page and updating the \\\"Last Updated\\\" date.\\nContact Us\\nIf you have any questions about this Privacy Policy, please contact us at info@dimension.im.\",\"uq8EoS\":\"Last Updated: 03.10 2024\",\"zwWKhA\":\"Learn more\",\"5ZMh6R\":\"Link 𝕏\",\"yQE2r9\":\"Loading\",\"tIaGLR\":\"MASK Stake is a staking project designed specifically for long-term holders of the MASK token. By staking MASK tokens on the MASK Stake platform, users can earn rewards from collaborating projects.\",\"GVGvph\":\"MASK Stake provides a stable income opportunity for MASK token holders while contributing to the growth of the MASK ecosystem. Whether you are an individual investor or an institutional investor, MASK Stake offers a secure and reliable staking platform to maximize the returns on your MASK tokens.\",\"qX6qUv\":\"Mask.io\",\"4HtGBb\":\"MAX\",\"cxA5GF\":[\"No enough \",[\"tokenSymbol\"],\" to claim\"],\"mB7ZUU\":\"No matched pool found.\",\"gZY4YJ\":\"No users have participated in staking yet.\",\"mu1gVr\":\"Not started\",\"HBEp+T\":\"Not started yet.\",\"J6n7sl\":\"Ongoing\",\"NukQm3\":\"Persistent Cookies: These cookies are stored on your device until they reach their expiration date or are manually deleted. They are used to remember your preferences and provide a personalized experience when you revisit the website.\",\"fE6ylD\":\"Please connect first\",\"OKEQCB\":\"Pool Liquidity\",\"LcET2C\":\"Privacy Policy\",\"vERlcd\":\"Profile\",\"R+X/he\":\"Profile Name\",\"51X4ns\":\"Profile Name is required\",\"BEn/IE\":\"Session Cookies: These cookies exist only during your visit to the website and are automatically deleted when you close your browser. They are used to maintain session state, such as storing the contents of your shopping cart.\",\"F8o9TO\":\"Share of Pool\",\"STjV6R\":\"Show Avatar\",\"c+Fnce\":\"Sign\",\"K/47pT\":\"Since 2019 to Now ｜\",\"eLyGnD\":\"Sorry, you are in a restricted region\",\"JYKRJS\":\"Stake\",\"b3zmge\":\"Stake $MASK for Rewards\",\"dy48iv\":\"Stake Mask\",\"Ue6VMJ\":\"Stake MASK\",\"boscAO\":\"Stake MASK Now\",\"+0dq3K\":\"Stake Successfully\",\"z87rAT\":\"Staked MASK\",\"NE0dj/\":\"Staked MASK can be unstake after the campaign ends.\",\"KIGhPb\":\"Staking is not available in restricted countries, including the US, Canada, China, Iran, North Korea, the Syrian Arab Republic, Netherlands, Crimea, Malaysia, Bangladesh, Bolivia and Cuba.\",\"ZoLkCW\":\"Staking Ranking\",\"V6XiJn\":[\"Successfully claimed \",[\"0\"],\" Tokens.\"],\"9FnQFX\":\"Successfully staked MASK Tokens.\",\"YsQ+vx\":\"Successfully unstaked MASK Tokens.\",\"XYLcNv\":\"Support\",\"PSGjEt\":\"Terms of Use\",\"8X0GDU\":\"Terms of Use for Mask.io\\nAcceptance of Terms\\nBy accessing and using Mask.io , you agree to be bound by these Terms of Use (\\\"Terms\\\"). If you do not agree to these Terms, you may not use our services.\\nChanges to Terms\\nWe reserve the right to modify these Terms at any time. We will notify you of any changes by posting the new Terms on this page. Your continued use of our services after any changes indicates your acceptance of the new Terms.\\nUse of Our Services\\n[Your Website Name] allows you to stake Mask to receive other token rewards. You agree to use our services only for lawful purposes and in accordance with these Terms.\\nAccount Responsibilities\\nYou are responsible for maintaining the confidentiality of your account information and for all activities that occur under your account.\\nYou agree to notify us immediately of any unauthorized use of your account.\\nProhibited Activities\\nYou agree not to engage in any of the following prohibited activities:\\nUsing our services for any illegal purpose or in violation of any local, state, national, or international law.\\nInfringing on the intellectual property rights of others or the intellectual property rights of Mask.io.\\nAttempting to gain unauthorized access to our services or another user's account.\\nInterfering with the security-related features of our services.\\nIntellectual Property\\nAll content, trademarks, logos, and service marks displayed on Mask.io are our property or the property of other third parties. You are not permitted to use these Marks without our prior written consent or the consent of such third party which may own the Marks.\\nDisclaimer of Warranties\\nOur services are provided on an \\\"as is\\\" and \\\"as available\\\" basis. Mask.io expressly disclaims all warranties of any kind, whether express or implied, including, but not limited to, the implied warranties of merchantability, fitness for a particular purpose, and non-infringement.\\nLimitation of Liability\\n[Your Website Name] shall not be liable for any indirect, incidental, special, consequential, or punitive damages, including but not limited to, damages for loss of profits, goodwill, use, data, or other intangible losses, resulting from the use or the inability to use our services.\\nIndemnification\\nYou agree to indemnify and hold harmless Mask.io and its officers, directors, employees, and agents from any and all claims, damages, expenses, and liabilities arising out of your use of our services, your violation of these Terms, or your violation of the rights of a third party.\\nGoverning Law\\nThese Terms shall be governed and construed in accordance with the laws of the People's Republic of China, without regard to its conflict of law provisions.\\nContact Us\\nIf you have any questions about these Terms, please contact us at info@dimension.im.\",\"rWQlhT\":\"The staking addresses need to pass Go+ security check. Note that staking is not available in some restricted regions. <0>More</0>\",\"Pim9Ok\":\"Third-Party Cookies: The Site may use cookies from third-party service providers for traffic analysis and user behavior tracking, as well as to display ads relevant to your interests.\",\"byuY4L\":\"This is the first phase of the MASK staking event. You will be able to retrieve your staked MASK tokens and reward tokens only after the event concludes.\",\"DYT0zx\":\"This website (hereinafter referred to as \\\"the Site\\\") uses cookies. By using the Site, you consent to the use of cookies as described in this Cookie Policy.\",\"sQlmlA\":\"This website is specifically created for $MASK investors and believers, where you can stake $MASK to earn token rewards!\",\"fPKzYp\":[\"Time \",[\"0\"]],\"C8gHY0\":\"To finish connecting, you must sign a message in your wallet to verify that you are the owner of this account.\",\"a7kfrG\":\"Total MASK Staked\",\"yxKbv1\":\"Total points\",\"qmYDVk\":\"Total Rewards\",\"mCJxHe\":\"Transaction failed.\",\"WabijI\":\"Transaction submitted!\",\"KTFOT8\":\"Types of Cookies We Use\",\"dMA7MR\":\"Unlock MASK\",\"vjn5CW\":\"Unlocking\",\"OFjZGo\":\"Unstake\",\"nnMoqX\":\"Unstake MASK\",\"vVo7n+\":\"Unstake MASK Time\",\"y/NCmf\":\"Unstake Successfully\",\"ilbzKn\":\"Updates to the Cookie Policy\",\"ik0mgO\":\"Verify your account\",\"L1QCpm\":\"Waiting for transaction confirmation\",\"9dsM+S\":\"We reserve the right to update this Cookie Policy at any time. The updated Cookie Policy will be posted on this page and will become effective immediately. We recommend checking back regularly for updates.\",\"a1Cyi2\":\"We use our own cookies.\",\"Qgqys4\":\"What Are Cookies?\",\"IwI0rY\":\"Wrong network\",\"hfcM4B\":\"You can claim after the event ends.\",\"2z+KTh\":\"You can manage and delete cookies through your browser settings. Please note that disabling cookies may affect your access to and experience on the Site.\",\"X1wp7N\":\"You have already claimed\",\"IV8aIx\":\"You have already unstaked\",\"I3z3fj\":[\"You have successfully claimed \",[\"0\"],\" \",[\"tokenSymbol\"],\".\"],\"fvkpR8\":[\"You have successfully staked \",[\"rawAmount\"],\" MASK.\"],\"H0XoEv\":[\"You have successfully unstaked \",[\"0\"],\" MASK\"],\"gGN2Rp\":\"Your Portal To The New, Open Internet.\",\"wmJARs\":\"Your wallet address is on the blacklist\",\"gWcGl/\":\"Your wallet cancelled the transaction.\"}");
